refactor: bump rockets version and adjust module implementations
